update no_warning_call utility in tests
